json changes

+++ commands.js
@@ -9,6 +9,7 @@ const { RichEmbed } = require('discord.js');
 const config = require("./config.json");
 const dscmds = require("./DSCMDS.js");
 
+const gdbfix = require("./gdbfix.js");
 const usrcmds = require("./users.js");
 const guildscmds = require("./modules/botcommands.js");
 const msglog = require("./modules/msglogger.js");
@@ -20,9 +21,12 @@ var gsetfile = "./guilds.json";
 const gset = require(gsetfile);
 const gsetsync = fs.readFileSync(gsetfile);
 
+var usetfile = "./users.json";
+const uset = require(usetfile);
+
 const insta = require("./instagram.js");
 
-module.exports = (message, client) => {
+module.exports =(message, client) => {
   
   //get arguments from the command
   const args = message.content.slice(config.prefix.length).split(' ');
@@ -76,8 +80,64 @@ module.exports = (message, client) => {
   
   if(message.content.startsWith(config.prefix + "pubgroom"))
   {
-    
+    /*
+    async function getFirstUser() {
+    let users = await getUsers();
+    return users[0].name;
+    */
+    async function getFirstUser() {
+    var bar = client.fetchUser("366565330719473667").then(user => {
+      return user.username;
+        })
+    return bar;
+    }  
+    console.log(getFirstUser());
   }    
+  
+  if(message.content.startsWith(config.prefix + "leaderboard"))
+  {
+    
+    const list = client.guilds.get(message.guild.id);
+    var a = [];
+    list.members.forEach(member => {
+      if(!uset[member.user.id])
+        uset[member.user.id]={}
+      if(!uset[member.user.id][message.guild.id])
+        uset[member.user.id][message.guild.id]={}
+      if(!uset[member.user.id][message.guild.id].points)
+        uset[member.user.id][message.guild.id].points=0;
+      fs.writeFile(usetfile, JSON.stringify(uset), (err) => console.error);
+      for(var i=0;i<10;i++)
+      {
+        var user=a[i];
+        if(a[i]==null)
+        {
+          a[i]=member.user.id;
+          break;
+        }
+        else if(uset[member.user.id][message.guild.id].points>uset[a[i]][message.guild.id].points)
+        {
+          for(var j=10;j<i;j--)
+          {
+            a[j]=a[j-1];
+          }
+          a[i]=member.user.id;
+          break;
+        }
+      }
+    });
+    var lb1="";
+    for(var i=0;i<10;i++)
+    {
+      if(!a[i])
+        break;
+      let user = client.users.get(a[i]);
+      var lb="";
+      lb=i+1 +". "+ user.username +" : "+uset[a[i]][message.guild.id].points + "points\n";
+      lb1=lb1+lb;
+    }
+    message.channel.send(lb1);
+  }
 //----------------------------server commands------------------------------------------------------
   if(message.content.startsWith(config.prefix + "server"))
   {
@@ -88,13 +148,29 @@ module.exports = (message, client) => {
  
 //----------------------------normal memeber command limiter---------------------------------------
 //----------------------------admin level commands-------------------------------------------------
-  
-  if(message.content.startsWith(config.prefix + "bd"))
+  const ownerarr = config.owner;
+  var oa=0;  
+  ownerarr.forEach(function(element) {
+
+    if(message.author.id == element)
+      oa=1; 
+  });
+  if((message.member.hasPermission("MANAGE_GUILD")) || (oa))
   {
-    var tot= args[0]+1;
-    message.channel.bulkDelete(tot);
+    if(message.content.startsWith(config.prefix + "bd"))
+    {
+      var tot= args[0]+1;
+      message.channel.bulkDelete(tot);
+    }
+    if(message.content.startsWith(config.prefix + "working"))
+    {
+      console.log("Working");
+    }
+    if(message.content.startsWith(config.prefix + "gdbfix"))
+    {
+      gdbfix(client, message);
+    }
   }
-  
 //----------------------------test commands--------------------------------------------------------
   if(message.content.startsWith(config.prefix + "check"))
   {
@@ -115,7 +191,7 @@ module.exports = (message, client) => {
   
   if(message.content.startsWith(config.prefix + "test"))
   {
-    https.get('https://www.googleapis.com/youtube/v3/channels?part=statistics&id=UC51OKFGqiMGco1QjZQI6gOg&key=AIzaSyDLQTe5eh8gUUeMQFueWeK5psOoino4458', (resp) => {
+    https.get('https://www.googleapis.com/youtube/v3/channels?part=statistics&id=UC51OKFGqiMGco1QjZQI6gOg&key='+ process.env.YTKEY+'', (resp) => {
   // A chunk of data has been recieved.
       resp.on('data', (chunk) => {
         chunk = JSON.parse(chunk);
@@ -129,14 +205,6 @@ module.exports = (message, client) => {
     insta(message,client);
   };
   
-  if(message.content.startsWith(config.prefix + "hii"))
-  {
-    console.log("test initiated");
-    setInterval(() => {
-    message.channel.send("Hii");
-    console.log("test successful");
-    }, 1800000);
-  };
 };
